+++ backend/src/main/java/cotato/backend/api/PostController.java
@@ -1,6 +1,6 @@
 package cotato.backend.api;
 
-import cotato.backend.domains.post.PostService;
+import cotato.backend.domains.post.service.PostService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import cotato.backend.common.dto.DataResponse;
-import cotato.backend.domains.post.dto.request.SavePostsByExcelRequest;
+import cotato.backend.api.dto.request.SavePostsByExcelRequest;
 import lombok.RequiredArgsConstructor;
 
 @RestController
